Добавил сервисы

+++ src/main/java/org/example/citydatabase/service/PersonServiceImpl.java
@@ -2,48 +2,54 @@ package org.example.citydatabase.service;
 
 import lombok.RequiredArgsConstructor;
 import org.example.citydatabase.dto.AddPersonRequestDto;
-import org.example.citydatabase.entity.*;
+import org.example.citydatabase.entity.Car;
+import org.example.citydatabase.entity.House;
+import org.example.citydatabase.entity.Passport;
+import org.example.citydatabase.entity.Person;
 import org.example.citydatabase.mapper.PersonMapper;
-import org.example.citydatabase.repository.*;
+import org.example.citydatabase.repository.PassportRepository;
+import org.example.citydatabase.repository.PersonRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Isolation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class PersonServiceImpl implements PersonService {
 
     private final PersonRepository personRepository;
-    private final HouseRepository houseRepository;
-    private final PersonHouseRepository personHouseRepository;
-    private final PassportRepository passportRepository;
-    private final CarRepository carRepository;
+    private final HouseService houseService;
+    private final PersonHouseService personHouseService;
+    private final PassportService passportService;
+    private final CarService carService;
     private final PersonMapper mapper;
 
     @Override
     public Person getPerson(Long personId) {
+        Optional<Person> optPerson = personRepository.findById(personId);
+        if (optPerson.isEmpty()) throw new NoSuchElementException("Person not found");
 
-        return personRepository.findById(personId).orElseThrow();
+        return optPerson.get();
     }
 
     @Override
     @Transactional(isolation = Isolation.REPEATABLE_READ)
     public Person addPerson(AddPersonRequestDto dto) {
         Person person = mapper.personFromAddPersonRequestDto(dto);
-        Passport passport = mapper.passportFromAddPersonRequestDto(dto);
+        Passport passport = passportService.addPassport();
+        person.setPassport(passport);
 
         personRepository.save(person);
-        passportRepository.save(passport);
 
         for (Long houseId : dto.getHousesId()) {
-            House house = houseRepository.findById(houseId).orElseThrow();
+            House house = houseService.getHouse(houseId);
+            if (house == null) throw new NoSuchElementException("House not found");
 
-            PersonHouse personHouse = new PersonHouse();
-            personHouse.setPerson(person);
-            personHouse.setHouse(house);
-            personHouseRepository.save(personHouse);
+            personHouseService.addPersonHouse(person, house);
         }
 
         return person;
@@ -52,13 +58,16 @@ public class PersonServiceImpl implements PersonService {
     @Override
     @Transactional(isolation = Isolation.REPEATABLE_READ)
     public void deletePerson(Long personId) {
-        Person person = personRepository.findById(personId).orElseThrow();
+        Optional<Person> optPerson = personRepository.findById(personId);
+        if (optPerson.isEmpty()) throw new NoSuchElementException("Person not found");
+
+        Person person = optPerson.get();
         Passport passport = person.getPassport();
         List<Car> cars = person.getCars();
 
-        personHouseRepository.deleteAllByPersonId(personId);
-        passportRepository.delete(passport);
-        carRepository.deleteAll(cars);
+        for (Car c : cars) carService.deleteCar(c.getId());
+        personHouseService.deleteAllByPersonId(personId);
+        passportService.deletePassport(passport.getId());
         personRepository.delete(person);
     }
 }
